Add new Wrap function and API
This does make most things cleaner and more logical: no more figuring out which
way the args go. The only weird bits are Fail() and Condition(). Also some of
the names may not make as much sense now.

I'm not 100% on the name of Must, but I'm not too worried about making something
that sounds like English. It should be possible to refactor the Must to be a
change of the function passed to the assertions, i.e. use Fatalf or Errorf.